+++ pkgs/test_api/lib/scaffolding.dart
@@ -17,7 +17,10 @@ export 'src/backend/configuration/tags.dart' show Tags;
 export 'src/backend/configuration/test_on.dart' show TestOn;
 export 'src/backend/configuration/timeout.dart' show Timeout;
 export 'src/scaffolding/spawn_hybrid.dart' show spawnHybridUri, spawnHybridCode;
-export 'src/scaffolding/test_structure.dart'
-    show group, test, setUp, setUpAll, tearDown, tearDownAll, addTearDown;
 export 'src/scaffolding/utils.dart'
-    show markTestSkipped, printOnFailure, pumpEventQueue, registerException;
+    show
+        addTearDown,
+        markTestSkipped,
+        printOnFailure,
+        pumpEventQueue,
+        registerException;