+++ pkgs/test_api/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.6.1-wip
+## 0.7.0-wip
 
 * Drop support for null unsafe Dart, bump SDK constraint to `3.0.0`.
 * Make some implementation classes `final`. These classes were never intended to
@@ -9,6 +9,9 @@
 * Mark an implementation class `interface`: `StackTraceMapper`.
 * Change the `Compiler` class into an `enum`.
 * Make `Fake` a `mixin class`.
+* Remove the definitions of `group`, `test`, `setUp`, `tearDown`, `setUpAll`,
+  and `tearDownAll`. These methods are available in `package:test_core` and
+  re-exported in `package:test`.
 
 
 ## 0.6.0